Issue #1986112, #1900918, #1990980 and #1973880 by riskogab, kepesv: Mega patch, fixes 4 issues and adds some new functionality
https://drupal.org/node/1900918
https://drupal.org/node/1990980
https://drupal.org/node/1973880
https://drupal.org/node/1986112

we've created a separated 'image annotator popup' module for this task. It provides a display formatter: http://screencast.com/t/dltamWovG and uses the query_qtip. If javascript is enabled on client side it puts the markers with bubbles on the image. If js not available it keeps the original way: annotations shows under the image.